lab5 test on vm

+++ lab5/04_random-function.sql
@@ -1,4 +1,4 @@
 SELECT 
-    ROUND(RANDOM() * 1000, 2) AS random_value
+    ROUND(RANDOM() * 1000::numeric, 2) AS random_value
 FROM accounts
 WHERE RANDOM() < 0.5;
